Split Dashboard container into 5 subcontainers
