+++ app/app.config.js
@@ -4,6 +4,8 @@
  * license that can be found in the LICENSE file.
  */
 
+import BigNumber from 'bignumber.js';
+
 const DEFAULT_REF_CODE = '00000000';
 
 export const MAIN_NET_GENESIS_BLOCK = '0xd4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3';
@@ -186,6 +188,8 @@ export const ABI_PROXY = [{ constant: false, inputs: [{ name: '_newOwner', type:
 export const ABI_TABLE_FACTORY = [{ constant: false, inputs: [{ name: '_newOwner', type: 'address' }], name: 'transfer', outputs: [], payable: false, type: 'function' }, { constant: true, inputs: [], name: 'getTables', outputs: [{ name: '', type: 'address[]' }], payable: false, type: 'function' }, { constant: true, inputs: [{ name: '_addr', type: 'address' }], name: 'isOwner', outputs: [{ name: '', type: 'bool' }], payable: false, type: 'function' }, { constant: true, inputs: [{ name: '', type: 'uint256' }], name: 'tables', outputs: [{ name: '', type: 'address' }], payable: false, type: 'function' }, { constant: true, inputs: [], name: 'owner', outputs: [{ name: '', type: 'address' }], payable: false, type: 'function' }, { constant: false, inputs: [{ name: '_smallBlind', type: 'uint96' }, { name: '_seats', type: 'uint256' }], name: 'create', outputs: [{ name: '', type: 'address' }], payable: false, type: 'function' }, { constant: true, inputs: [], name: 'tokenAddress', outputs: [{ name: '', type: 'address' }], payable: false, type: 'function' }, { constant: true, inputs: [], name: 'oracleAddress', outputs: [{ name: '', type: 'address' }], payable: false, type: 'function' }, { constant: false, inputs: [{ name: '_token', type: 'address' }, { name: '_oracle', type: 'address' }], name: 'configure', outputs: [], payable: false, type: 'function' }];
 
 export const TIMEOUT_PERIOD = (state) => (state === 'waiting' || state === 'dealing') ? 10 : 40;
+
+export const ETH_FISH_LIMIT = new BigNumber(0.1);
 
 function bn(num) {
   return num * 1000000000000;
