+++ app/containers/Dashboard/constants.js
@@ -1,0 +1,2 @@
+export const LOOK_BEHIND_PERIOD = 4 * 60 * 24;
+
